ocf: Provide dictionary-based access to typed properties
Code written with a specific resource type in mind will most naturally
want to use Python attributes in order to access resource properties,
with expressions such as "x.prop.temperature".

Code written to handle arbitrary resource types may need the ability
to iterate over available resource properties, and to access
properties using the OCF property names (which may not match the
Python property names due to reserved keywords or namespace
collisions).

Allow ResourceType classes and objects to act as dictionaries where
the keys are the OCF property names and the values are respectively
the Python property objects and the current state values.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>