ocf: Use canonicalisation when converting to or from raw state value
The raw state value as constructed by deserialising a CBOR or JSON
object may be missing type information.  For example: a UUID is likely
to be represented in JSON as a string.

When reading the property, the accessor must therefore be prepared to
convert the raw state value to the appropriate Python type
(e.g. uuid.UUID).

When serialising the raw state value, the type information may be
required.  For example: a UUID must be represented in CBOR using major
type 2 and tag 37.

It is viable to require that deserialisation preserves sufficient
information to allow for reserialisation using the same encoding.  For
example: a UUID deserialised from CBOR will be represented in the raw
state value as a uuid.UUID (since CBOR includes type information), and
so will be reserialised into CBOR as a UUID.  A UUID deserialised from
JSON will be represented in the raw state value as a string, and so
will be reserialised into JSON as a string.  Since JSON uses strings
to represent UUIDs, this is sufficient despite the absence of type
information in the raw state.

When writing the property, the intended serialisation may not be
known.  The accessor must therefore ensure that the new raw state
value preserves enough type information to allow for correct
serialisation in either encoding.

These requirements on the property accessors may be satisfied by
defining a "canonicalise" method which accepts any permitted raw state
representation (e.g. a string or a uuid.UUID, for a UUID property) and
returns the correctly typed Python object (e.g. a uuid.UUID).  This
canonicalisation can then be applied when either reading or writing
the property.

As a bonus, this method then provides a mechanism for canonicalising
the raw state value, thereby allowing a raw state deserialised from
JSON (and thus missing type information) to be reserialised correctly
to CBOR (which requires type information).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>