ocf: Provide only indirect access to raw state values
A Resource object may currently be treated as a dictionary providing
access to the raw state values.  This creates a situation in which
accidentally writing an expression such as "res['temperature']"
instead of "res.prop['temperature']" would result in code that
accesses the uncanonicalised raw state value.  This will often happen
to work but may introduce subtle bugs resulting in incorrect
serialisations.

Avoid this class of potential bugs by removing the ability for a
Resource object to act as a dictionary, and providing access to raw
state values only via the explicit "state" attribute.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>