ocf: Canonicalise "rt" and "if" using OrderedSet
At a protocol level "rt" and "if" are arrays, but at a conceptual
level they are better modelled as sets since duplicate values are
meaningless.  Use OrderedSet as the canonical representation for these
properties, to provide an appropriate conceptual model while
minimising surprising ordering changes at the protocol level.

Note that there is a bug in the OrderedSet __ge__ and __gt__
comparison methods that causes false positives: work around this bug
by using "<=" instead of ">=" in our __subclasscheck__ and
__instancecheck__ methods.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>