ocf: Use empty containers as default value for container properties
Allow Property subclasses to provide default values if not explicitly
overridden, and use empty containers as the default values for
ArrayProperty and OrderedSetProperty.  This is likely to simplify most
use cases of container-valued properties.

Define an "Unspecified" singleton so that Property instances may
specify an explicit default value of None if needed.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>